Print current weather on monitor app

+++ src/main/java/dev/muscaw/monitor/weather/domain/RelativeHumidity.java
@@ -1,3 +1,3 @@
 package dev.muscaw.monitor.weather.domain;
 
-public record RelativeHumidity(float percentage) {}
+public record RelativeHumidity(int percentage) {}
